% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hash_lemma.R
\docType{data}
\name{hash_lemma_ro}
\alias{hash_lemma_ro}
\title{Romanian Lemma Dictionary}
\format{A data frame with 296754 rows and 2 variables}
\usage{
data(hash_lemma_ro)
}
\description{
A dataset containing Romanian tokens and corresponding lemmas. This data set
comes from \url{http://www.lexiconista.com}, and is made freely available
under the ODbL 1.0 license (\url{https://opendatacommons.org/licenses/odbl/summary}).
Note that tokes with more than one lemma were dropped from the data set.
}
\details{
\itemize{
  \item token. An inflected token with affixes
  \item lemma. A base form
}
}
\references{
\url{http://www.lexiconista.com}
}
\keyword{datasets}

