% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemma_lang.R
\name{lemmatize_ast}
\alias{lemmatize_ast}
\alias{lemmatize_bg}
\alias{lemmatize_ca}
\alias{lemmatize_cs}
\alias{lemmatize_en}
\alias{lemmatize_et}
\alias{lemmatize_fr}
\alias{lemmatize_gl}
\alias{lemmatize_de}
\alias{lemmatize_hu}
\alias{lemmatize_ga}
\alias{lemmatize_gv}
\alias{lemmatize_it}
\alias{lemmatize_fa}
\alias{lemmatize_pl}
\alias{lemmatize_pt}
\alias{lemmatize_ro}
\alias{lemmatize_gd}
\alias{lemmatize_sk}
\alias{lemmatize_sl}
\alias{lemmatize_es}
\alias{lemmatize_sv}
\alias{lemmatize_uk}
\alias{lemmatize_cy}
\title{Asturian Lemmatizer}
\usage{
lemmatize_ast(x, dictionary = lemmar::hash_lemma_ast, ...)

lemmatize_bg(x, dictionary = lemmar::hash_lemma_bg, ...)

lemmatize_ca(x, dictionary = lemmar::hash_lemma_ca, ...)

lemmatize_cs(x, dictionary = lemmar::hash_lemma_cs, ...)

lemmatize_en(x, dictionary = lemmar::hash_lemma_en, ...)

lemmatize_et(x, dictionary = lemmar::hash_lemma_et, ...)

lemmatize_fr(x, dictionary = lemmar::hash_lemma_fr, ...)

lemmatize_gl(x, dictionary = lemmar::hash_lemma_gl, ...)

lemmatize_de(x, dictionary = lemmar::hash_lemma_de, ...)

lemmatize_hu(x, dictionary = lemmar::hash_lemma_hu, ...)

lemmatize_ga(x, dictionary = lemmar::hash_lemma_ga, ...)

lemmatize_gv(x, dictionary = lemmar::hash_lemma_gv, ...)

lemmatize_it(x, dictionary = lemmar::hash_lemma_it, ...)

lemmatize_fa(x, dictionary = lemmar::hash_lemma_fa, ...)

lemmatize_pl(x, dictionary = lemmar::hash_lemma_pl, ...)

lemmatize_pt(x, dictionary = lemmar::hash_lemma_pt, ...)

lemmatize_ro(x, dictionary = lemmar::hash_lemma_ro, ...)

lemmatize_gd(x, dictionary = lemmar::hash_lemma_gd, ...)

lemmatize_sk(x, dictionary = lemmar::hash_lemma_sk, ...)

lemmatize_sl(x, dictionary = lemmar::hash_lemma_sl, ...)

lemmatize_es(x, dictionary = lemmar::hash_lemma_es, ...)

lemmatize_sv(x, dictionary = lemmar::hash_lemma_sv, ...)

lemmatize_uk(x, dictionary = lemmar::hash_lemma_uk, ...)

lemmatize_cy(x, dictionary = lemmar::hash_lemma_cy, ...)
}
\arguments{
\item{x}{A vector of strings.}

\item{dictionary}{A dictionary of base terms and lemmas to use for
replacement.  The first column should be the full word form in lower case
while the second column is the corresponding replacement lemma.}

\item{\ldots}{Other arguments passed to \code{\link[textshape]{split_token}}.}
}
\description{
A wrapper for \code{lemmatize} that has the default
\code{dictionary = lemmar::hash_lemma_ast} set.

A wrapper for \code{lemmatize} that has the default
\code{dictionary = lemmar::hash_lemma_bg} set.

A wrapper for \code{lemmatize} that has the default
\code{dictionary = lemmar::hash_lemma_ca} set.

A wrapper for \code{lemmatize} that has the default
\code{dictionary = lemmar::hash_lemma_cs} set.

A wrapper for \code{lemmatize} that has the default
\code{dictionary = lemmar::hash_lemma_en} set.

A wrapper for \code{lemmatize} that has the default
\code{dictionary = lemmar::hash_lemma_et} set.

A wrapper for \code{lemmatize} that has the default
\code{dictionary = lemmar::hash_lemma_fr} set.

A wrapper for \code{lemmatize} that has the default
\code{dictionary = lemmar::hash_lemma_gl} set.

A wrapper for \code{lemmatize} that has the default
\code{dictionary = lemmar::hash_lemma_de} set.

A wrapper for \code{lemmatize} that has the default
\code{dictionary = lemmar::hash_lemma_hu} set.

A wrapper for \code{lemmatize} that has the default
\code{dictionary = lemmar::hash_lemma_ga} set.

A wrapper for \code{lemmatize} that has the default
\code{dictionary = lemmar::hash_lemma_gv} set.

A wrapper for \code{lemmatize} that has the default
\code{dictionary = lemmar::hash_lemma_it} set.

A wrapper for \code{lemmatize} that has the default
\code{dictionary = lemmar::hash_lemma_fa} set.

A wrapper for \code{lemmatize} that has the default
\code{dictionary = lemmar::hash_lemma_pl} set.

A wrapper for \code{lemmatize} that has the default
\code{dictionary = lemmar::hash_lemma_pt} set.

A wrapper for \code{lemmatize} that has the default
\code{dictionary = lemmar::hash_lemma_ro} set.

A wrapper for \code{lemmatize} that has the default
\code{dictionary = lemmar::hash_lemma_gd} set.

A wrapper for \code{lemmatize} that has the default
\code{dictionary = lemmar::hash_lemma_sk} set.

A wrapper for \code{lemmatize} that has the default
\code{dictionary = lemmar::hash_lemma_sl} set.

A wrapper for \code{lemmatize} that has the default
\code{dictionary = lemmar::hash_lemma_es} set.

A wrapper for \code{lemmatize} that has the default
\code{dictionary = lemmar::hash_lemma_sv} set.

A wrapper for \code{lemmatize} that has the default
\code{dictionary = lemmar::hash_lemma_uk} set.

A wrapper for \code{lemmatize} that has the default
\code{dictionary = lemmar::hash_lemma_cy} set.
}
