% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemmatize.R
\name{lemmatize}
\alias{lemmatize}
\title{Lemmatize a Vector of Strings}
\usage{
lemmatize(x, dictionary = lemmar::hash_lemma_en, ...)
}
\arguments{
\item{x}{A vector of strings.}

\item{dictionary}{A dictionary of base terms and lemmas to use for
replacement.  The first column should be the full word form in lower case
while the second column is the corresponding replacement lemma.}

\item{\ldots}{Other arguments passed to \code{\link[textshape]{split_token}}.}
}
\value{
Returns a vector of lemmatized strings.
}
\description{
Lemmatize a vector of strings.
}
\note{
The lemmatizer splits the string apart into tokens for speed
optimization.  After the lemmatizing occurs the strings are pasted back
together.  The strings are not guaranteed to retain exact spacing of the
original.
}
\examples{
x <- c(
    'the dirtier dog has eaten the pies',
    'that shameful pooch is tricky and sneaky',
    "He opened and then reopened the food bag",
    'There are skies of blue and red roses too!',
    NA,
    "The doggies, well they aren't joyfully running.",
    "The daddies are coming over...",
    "This is 34.546 above"
)

lemmatize(x)

## Bigger data set
library(dplyr)
presidential_debates_2012$dialogue \%>\%
    head()
gc(); tic <- Sys.time()
presidential_debates_2012$dialogue \%>\%
    lemmatize() \%>\%
    head()
cat(sprintf(
    '\%s seconds for \%s rows of text\\n',
    round(Sys.time() - tic, 2),
    nrow(presidential_debates_2012)
))
}
