% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lemma-package.R
\docType{package}
\name{lemmar}
\alias{lemmar}
\alias{lemmar-package}
\alias{package-lemmar}
\title{Dictionary Based Lemmatization}
\description{
Utilizes tokenization and dictionary lookup for lemmatization of text.
Lemmatization is defined as  "grouping together the inflected forms of a word
so they can be analysed as a single item" (wikipedia).  While dictionary
lookup of tokens is not a true morphological analysis, this style of lemma
replacement is fast and typically still robust for many applications
}

